refactor: utils에서 dom관리하도록 수정
